+++ changelog.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support for Heater on/off command
 - Added support for setting pool and spa setpoints
 - Added favicon.ico because the 404 errors were annoying me
+- Added support for static files such as css/js etc.
 
 ### Removed